added more icons to the library #315
I added more icons to the component

`man.svg` also had a big file. So i remaked it. I also made it all blue cause the background is transparent en some of the content wouldn't be clear. It's a contrast issue. We have to look if this is fine when we talk with the owner