implementing filter on `tekenmodus` page.
Here you can find the code to filter on the page. @Robin1224 only said that this isn't the proper way to do this. You want the url included with the checkbox. So I have to look into that. On svelte 5 this progress. If the update is through I will improve this. @DamianR2004 will update this first