applying a proper overview
for #423 I saw that these images weren't loading in right. I defined the width and height here to prevent a layout-shift. I gave the types with it, cause the W3C said it was better to do so. and I made the images load in with lazyloading to make the page load faster